requirements.txt
Django>=4.2,<5
djangorestframework>=3.14
django-cors-headers>=4.3
djangorestframework-simplejwt>=5.3
dj-database-url>=2.1
psycopg2-binary>=2.9
Pillow>=10.0
reportlab>=4.0
